pyproject.toml
[project]
name = "lg-discord-bot"
version = "0.1.0"
description = "LangGraph ReAct agent with Discord bot interface, powered by Gemini"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=0.2.14",
    "langchain-google-genai>=4.2.1",
    "langchain-tavily>=0.1",
    "discord.py>=2.7.1",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.1",
    "python-ulid>=3.0.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]

[tool.setuptools.package-dir]
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[project.scripts]
discord-bot = "bot.discord_bot:main"
telegram-bot = "bot.telegram_bot:main"
